added multiple pipes, very bad done but will be improved at a later stage

+++ src/model/Bird.java
@@ -45,7 +45,7 @@ class Bird implements IPositionable {
   }
 
   public void reset() {
-    this.position = new Point2D.Double(50, 280);
+    this.position = new Point2D.Float(50, 280);
   }
 
   @Override
